% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{use_skeletons}
\alias{use_skeletons}
\title{Add skeletons dependencies to the given tag}
\usage{
use_skeletons(tag)
}
\arguments{
\item{tag}{Any html tag. In practice, best use it to attach dependencies
to the main UI wrapper.}
}
\value{
The original tag with all necessary skeletons dependencies
}
\description{
Add skeletons dependencies to the given tag
}
